post методы в регистрации принимают параметры в теле json
